backend/pyproject.toml
[tool.poetry]
name = "sigil-backend"
version = "0.1.0"
description = "Django backend for SigilApp"
authors = ["Jason Thorgny Martineau"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12, <4.0"
django = ">=6.0.1,<7.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
djoser = ">=2.3.3,<3.0.0"
pillow = "^12.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
